feat: add musicbrainz service
