+++ sidetrack/enrichment/mb.py
@@ -1,10 +1,10 @@
 from __future__ import annotations
 
-import asyncio
 import logging
-import time
 
 import httpx
+
+from sidetrack.services.musicbrainz import MusicBrainzService
 
 from . import TrackRef
 
@@ -12,11 +12,10 @@ from . import TrackRef
 class MusicBrainzAdapter:
     """Simple wrapper for MusicBrainz lookups."""
 
-    api_root = "https://musicbrainz.org/ws/2"
-
     def __init__(self, client: httpx.AsyncClient | None = None) -> None:
         headers = {"User-Agent": "SideTrack/0.1 (+https://example.com)"}
         self._client = client or httpx.AsyncClient(headers=headers)
+        self._service = MusicBrainzService(self._client)
 
     logger = logging.getLogger(__name__)
 
@@ -26,31 +25,15 @@ class MusicBrainzAdapter:
     async def recording_by_isrc(self, isrc: str) -> TrackRef | None:
         """Return basic metadata for a recording identified by its ISRC."""
 
-        url = f"{self.api_root}/isrc/{isrc}"
-        params = {"inc": "recordings+artist-credits", "fmt": "json"}
-        start = time.perf_counter()
         try:
-            resp = await self._client.get(url, params=params, timeout=30)
-            resp.raise_for_status()
-            data = resp.json()
-            recording = (data.get("recordings") or [{}])[0]
-            title = recording.get("title") or ""
-            artists = [
-                c.get("artist", {}).get("name")
-                for c in recording.get("artist-credit", [])
-                if c.get("artist", {}).get("name")
-            ]
-            mbid = recording.get("id")
-            duration = time.perf_counter() - start
-            self.logger.info("enrichment_mb_success isrc=%s duration=%.3fs", isrc, duration)
+            data = await self._service.recording_by_isrc(isrc)
+            if not data.get("recording_mbid"):
+                return None
+            title = data.get("title") or ""
+            artists = data.get("artists") or []
+            mbid = data.get("recording_mbid")
+            self.logger.info("enrichment_mb_success isrc=%s", isrc)
             return TrackRef(title=title, artists=artists, isrc=isrc, lastfm_mbid=mbid)
-        except Exception as exc:
-            duration = time.perf_counter() - start
-            self.logger.warning(
-                "enrichment_mb_fail isrc=%s duration=%.3fs error=%s", isrc, duration, str(exc)
-            )
+        except Exception as exc:  # pragma: no cover - error logging
+            self.logger.warning("enrichment_mb_fail isrc=%s error=%s", isrc, str(exc))
             raise
-        finally:
-            # Honour MusicBrainz rate-limiting guidelines
-            await asyncio.sleep(1)
-            self.logger.info("mb_rate_limit_sleep seconds=1.0")
